fix examples and test
