+++ fedot_ind/api/utils/api_init.py
@@ -82,9 +82,9 @@ class ApiManager:
         self.config['optimizer'] = kwargs.get('optimizer', IndustrialEvoOptimizer)
         self.config['initial_assumption'] = kwargs.get('initial_assumption', None)
         if self.config['initial_assumption'] is None:
-            self.config['initial_assumption'] = Either(value=self.strategy,
+            self.config['initial_assumption'] = Either(value=self.strategy_class,
                                                        monoid=[self.preset,
-                                                               self.strategy == 'anomaly_detection']). \
+                                                               self.strategy_class == 'anomaly_detection']). \
                 either(left_function=fedot_init_assumptions,
                        right_function=fedot_init_assumptions)
