+++ fedot_ind/api/utils/checkers_collections.py
@@ -117,9 +117,9 @@ class DataCheck:
                 len(
                     data_list[0])))
 
-        have_predict_horizon = Either(value=False, monoid=[True, len(self.industrial_task_params) == 0]).either(
+        have_predict_horizon = Either(value=False, monoid=[True, self.industrial_task_params is None]).either(
             left_function=lambda l: self.industrial_task_params['data_type'] == 'time_series' and
-            'detection_window' in self.industrial_task_params.keys(),
+                                    'detection_window' in self.industrial_task_params.keys(),
             right_function=lambda r: r)
 
         task = Either(
